fix: Allow VWS handle UNC path
Silver Searcher does not return //xxx or \\xxx as file path if the
string concludes with a slash. It strips the first slash, which Vim
reads as a relative path, rather than the UNC path it should have been.
Stip the trailing slash, and Ag will search with \\xxx format UNC.